requirements.txt
fastapi>=0.104
uvicorn[standard]>=0.24
pydantic-settings>=2.0
azure-storage-blob>=12.0
rtree>=1.1
